Renamed MakeInvitation to CreateInvitation and moved GetInvitationDetails to Queries folder
