+++ Application/Commands/Invitations/CreateInvitation/CreateInvitationRequest.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+
+namespace Application.Commands.Invitations.CreateInvitation
+{
+    public record CreateInvitationRequest : IRequest<string>
+    {
+        public string ServerId { get; set; }
+        public bool IncludeUser { get; set; }
+        public DateTime? ExpireTime { get; set; }
+    }
+}
