--- Application/Commands/Invitations/MakeInvitation/MakeInvitationRequest.cs
@@ -1,12 +1,0 @@
-﻿using Application.Models;
-using MediatR;
-
-namespace Application.Commands.Invitations.MakeInvitation
-{
-    public record MakeInvitationRequest : IRequest<string>
-    {
-        public string ServerId { get; set; }
-        public bool IncludeUser { get; set; }
-        public DateTime? ExpireTime { get; set; }
-    }
-}
