+++ Application/Queries/GetInvitationDetails/GetInvitationDetailsRequest.cs
@@ -1,11 +1,14 @@
-﻿using System.ComponentModel;
+﻿using MediatR;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
-using MediatR;
 
-namespace Application.Commands.Invitations.GetInvitationDetails
+namespace Application.Queries.GetInvitationDetails
 {
     public record GetInvitationDetailsRequest : IRequest<InvitationDetailsDto>
     {
+        /// <summary>
+        /// The unique identifier of the invitation for which to retrieve details
+        /// </summary>
         [Required]
         [StringLength(24, MinimumLength = 24)]
         [DefaultValue("5f95a3c3d0ddad0017ea9291")]
